pyproject.toml
# SPDX-FileCopyrightText: 2025 Daniel Eder
#
# SPDX-License-Identifier: MIT

[project]
name = "pdf-outliner"
version = "0.1.0"
description = "AI-powered PDF outliner that adds bookmarks based on detected headings"
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }
authors = [
    { name = "Daniel Eder", email = "danieleder@outlook.at" }
]
dependencies = [
    "litellm>=1.0.0",
    "pypdf>=3.17.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/daniel-eder/pdf-outliner"
Repository = "https://github.com/daniel-eder/pdf-outliner"
Issues = "https://github.com/daniel-eder/pdf-outliner/issues"

[project.scripts]
pdf-outliner = "pdf_outliner:main"

[build-system]
requires = ["uv_build>=0.8.24,<0.9.0"]
build-backend = "uv_build"
